+++ textures/texture.cu
@@ -4,7 +4,7 @@
 #include <textures/solid_texture.h>
 #include <textures/texture.h>
 
-__device__ color texture_color_at(struct Texture* texture, double u, double v, vec3& point)
+__device__ color texture_color_at(struct Texture* texture, float u, float v, vec3& point)
 {
     switch (texture->texture_type) {
     case SOLID_COLOR:
